requirements.txt
python-telegram-bot==20.7
python-dotenv
aiohttp==3.9.1
sqlalchemy
psycopg2-binary
requests
cachetools
dependency-injector==4.41.0
pydantic==2.5.2
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1 